refactor: LineStation 도메인명 변경, 패키지 분리
- Line과 Station을 연결해주는 의미에서 두 도메인을 합친 이름으로 진행했으나,
쉬운 단어가 존재해서 section 으로 변경
- 각 도메인별로 페키지를 분리했습니다.
