--- src/main/java/nextstep/subway/domain/Line.java
@@ -1,63 +1,0 @@
-package nextstep.subway.domain;
-
-import java.util.List;
-import javax.persistence.Column;
-import javax.persistence.Embedded;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import nextstep.subway.dto.LineRequest;
-
-@Entity
-public class Line extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(unique = true)
-    private Long id;
-    @Column(unique = true, nullable = false)
-    private String name;
-    @Column(unique = true, nullable = false)
-    private String color;
-    @Embedded
-    private LineStations lineStations = LineStations.create();
-
-    protected Line() {
-    }
-
-    public Line(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public void addLineStation(LineStation lineStation) {
-        lineStations.add(lineStation);
-        lineStation.addLine(this);
-    }
-
-    public List<Station> getStations() {
-        return lineStations.getStations();
-    }
-
-    public Line of(LineRequest lineRequest) {
-        return new Line(lineRequest.getName(), lineRequest.getColor());
-    }
-
-    public void updateInfo(Line updateLine) {
-        this.name = updateLine.name;
-        this.color = updateLine.color;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-}
