--- src/main/java/nextstep/subway/dto/LineRequest.java
@@ -1,51 +1,0 @@
-package nextstep.subway.dto;
-
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.LineStation;
-import nextstep.subway.domain.Station;
-
-public class LineRequest {
-    private String name;
-    private String color;
-    private Long upStationId;
-    private Long downStationId;
-    private int distance;
-
-    protected LineRequest() {
-    }
-
-    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
-        this.name = name;
-        this.color = color;
-        this.upStationId = upStationId;
-        this.downStationId = downStationId;
-        this.distance = distance;
-    }
-    public Line toLine(Station upStation, Station downStation) {
-        Line line = new Line(name, color);
-        LineStation lineStation = new LineStation(upStation, downStation, distance);
-        line.addLineStation(lineStation);
-        return line;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-
-}
