--- src/main/java/nextstep/subway/dto/LineStationResponse.java
@@ -1,31 +1,0 @@
-package nextstep.subway.dto;
-
-public class LineStationResponse {
-    private Long lineId;
-    private Long upStationId;
-    private Long downStationId;
-    private int distance;
-
-    public LineStationResponse(Long lineId, Long upStationId, Long downStationId, int distance) {
-        this.lineId = lineId;
-        this.upStationId = upStationId;
-        this.downStationId = downStationId;
-        this.distance = distance;
-    }
-
-    public Long getLineId() {
-        return lineId;
-    }
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-}
