+++ src/main/java/nextstep/subway/line/dto/LineResponse.java
@@ -1,0 +1,51 @@
+package nextstep.subway.line.dto;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.dto.StationResponse;
+
+public class LineResponse {
+    private Long id;
+    private String name;
+    private String color;
+    private List<StationResponse> stations;
+
+    public LineResponse() {
+    }
+
+    public LineResponse(Long id, String name, String color, List<Station> stations) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+        this.stations = stations.stream()
+                .map(StationResponse::of)
+                .collect(Collectors.toList());
+    }
+
+    public static LineResponse of(Line line) {
+        return new LineResponse(
+                line.getId(),
+                line.getName(),
+                line.getColor(),
+                line.getStations()
+        );
+    }
+
+    public Long getId() {
+        return this.id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
+    }
+}
