+++ src/main/java/nextstep/subway/section/dto/SectionResponse.java
@@ -1,0 +1,31 @@
+package nextstep.subway.section.dto;
+
+public class SectionResponse {
+    private Long lineId;
+    private Long upStationId;
+    private Long downStationId;
+    private int distance;
+
+    public SectionResponse(Long lineId, Long upStationId, Long downStationId, int distance) {
+        this.lineId = lineId;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public Long getLineId() {
+        return lineId;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
