+++ src/main/java/nextstep/subway/station/domain/Station.java
@@ -1,0 +1,28 @@
+package nextstep.subway.station.domain;
+
+import javax.persistence.*;
+import nextstep.subway.common.domain.BaseEntity;
+
+@Entity
+public class Station extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(unique = true)
+    private String name;
+
+    public Station() {
+    }
+
+    public Station(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
