+++ src/main/java/nextstep/subway/station/dto/StationRequest.java
@@ -1,0 +1,15 @@
+package nextstep.subway.station.dto;
+
+import nextstep.subway.station.domain.Station;
+
+public class StationRequest {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public Station toStation() {
+        return new Station(name);
+    }
+}
